Updated helpers (renamed to apiRequester) Co-authored-by: elena-vi <elena@madetech.com>
