+++ api/gateways/SearchPropertiesGateway.js
@@ -3,7 +3,7 @@ module.exports = makeGetRequest => {
   return async postcode => {
     var result;
     result = await makeGetRequest({
-      url: `/addresses?postcode=${postcode}`
+      url: `/addressesss?postcode=${postcode}`
     }).then(response => {
       return response.data;
     }).catch(error => {
